// src/constants/call.enum.ts
export enum CallTypeEnum {
  Disorientation = 'disorientation',
  Gossip = 'gossip',
  TreasureHunt = 'treasureHunt',
  Inferring = 'inferring',
  LackOfChoice = 'lackOfChoice',
  Deprivation = 'deprivation',
  RequestOfHelp = 'requestOfHelp',
  Temptation = 'temptation',
  Synchronism = 'synchronism',
  PushToAction = 'pushToAction',
  HeraldOfChange = 'heraldOfChange',
}